[OPERATOR-845] Force ccm upgrade during px upgrade, and get PX version from release manifest env var as well
