+++ deploy/crds/core_v1_storagecluster_crd.yaml
@@ -926,7 +926,7 @@ spec:
                       image:
                         type: string
                         description: Docker image of the telemetry container.
-                      imageLogUpload:
+                      logUploaderImage:
                         type: string
                         description: Docker image of the log-upload-service container.
               security:
@@ -1472,7 +1472,7 @@ spec:
                     description: Desired image for metrics collector proxy.
                   telemetryProxy:
                     type: string
-                    description: Desired image for telemetry.
+                    description: Desired image for telemetry proxy.
                   logUploader:
                     type: string
                     description: Desired image for log uploader.
